requirements.txt
fastapi
uvicorn[standard]
python-multipart

pypdf
python-docx
pandas
openpyxl
numpy

tiktoken
nltk

sentence-transformers
torch
transformers

chromadb

openai
google-genai

python-dotenv
pydantic
scikit-learn

orjson
httpx
aiofiles

tabulate

pydantic-settings
rank_bm25

google-generativeai
later

# Evaluation
ragas
datasets
langchain-google-genai
langchain-community
pandas
